Revert "feat(all): Pr v3.5.2 batch1"

+++ master/node_selector.go
@@ -71,7 +71,7 @@ type Node interface {
 	PartitionCntLimited() bool
 	IsActiveNode() bool
 	IsWriteAble() bool
-	GetPartitionLimitCnt() uint64
+	GetPartitionLimitCnt() uint32
 	GetTotal() uint64
 	GetUsed() uint64
 	GetAvailableSpace() uint64
